Refactor Cleaner Mapping
